Fix the atime issue when finding exclude files
To delete the excluded files from the snapshot, the snapshot must be
writable. In order to find the files to be deleted from the snapshot we
need to walk through it. Therefore we change the `atime` attribute of
all the files as the snapshot is writable and thus changing the file and
causing a new copy for some reason.

Solution: We make the snapshot readonly, walk through it to collect the
files to be deleted in a list. Then we change the snapshot to readwrite
and delete al the files on the list. Then we set the snapshot back to
readonly.